chore: rename aigne hub
